permission and location fixed

+++ download_phytoMorph.sh~
@@ -64,15 +64,39 @@ sudo chmod +x $4/phytoG.png
 echo "Done download: phytoG.png "
 ################################################################################################################
 ################################################################################################################
+
 '
-sudo $4/ifExist.sh '-d' 'icommands.x86_64.tar.bz2' '' '/usr/sbin'
-sudo $4/ifExist.sh '-f' 'irodsFs' '/usr/sbin' 
-'
-sudo mv /usr/sbin/icommands/icd /usr/sbin/
-sudo wget -O icommands.x86_64.tar.bz2 http://davos.cyverse.org/irods-rest/rest/fileContents/iplant/home/nmiller/publicData/icommands.x86_64.tar.bz2?ticket=acamNrXKjPYRxtM
+sudo wget -O $4/icommands.x86_64.tar.bz2 http://davos.cyverse.org/irods-rest/rest/fileContents/iplant/home/nmiller/publicData/icommands.x86_64.tar.bz2?ticket=acamNrXKjPYRxtM
 sudo tar xjf icommands.x86_64.tar.bz2 -C /usr/sbin/
 sudo mv /usr/sbin/icommands/icd /usr/sbin/ 
 
 sudo wget -O /usr/sbin/irodsFs http://davos.cyverse.org/irods-rest/rest/fileContents/iplant/home/nmiller/publicData/irodsFs?ticket=WIfveh6JwMykqun
 sudo chmod +x /usr/sbin/irodsFs
+'
+
+
+
+myMCRtar="$4/icommands.x86_64.tar.bz2"
+if [ -f "$myMCRZtar" ]
+then
+	echo "$myMCRtar found. Skip download $myMCRtar"
+	sudo tar xjf icommands.x86_64.tar.bz2 -C /usr/sbin/
+	sudo mv /usr/sbin/icommands/icd /usr/sbin/
+else
+	echo "$myMCRtar not found."
+	sudo wget -O $4/icommands.x86_64.tar.bz2 http://davos.cyverse.org/irods-rest/rest/fileContents/iplant/home/nmiller/publicData/icommands.x86_64.tar.bz2?ticket=acamNrXKjPYRxtM
+	sudo tar xjf icommands.x86_64.tar.bz2 -C /usr/sbin/
+	sudo mv /usr/sbin/icommands/icd /usr/sbin/
+fi
+
+myFs="/usr/sbin/irodsFs"
+if [ -f "$myFs" ]
+then
+	echo "$myFs found. Skip download $myFs"
+	sudo chmod +x /usr/sbin/irodsFs
+else
+	echo "$myFs not found."
+	sudo wget -O /usr/sbin/irodsFs http://davos.cyverse.org/irods-rest/rest/fileContents/iplant/home/nmiller/publicData/irodsFs?ticket=WIfveh6JwMykqun
+	sudo chmod +x /usr/sbin/irodsFs
+fi
